convert return value of sec_to_ms to int

+++ src/utils/converters.py
@@ -1,2 +1,2 @@
 def sec_to_ms(sec: float) -> int:
-    return sec * 1000
+    return int(sec * 1000)
